# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Bonza Bike Tours REST API",
		"description": "REST API description for the Bonza Bike Tours BMS dashboard.",
		"contact": {
			"name": "API Support",
			"email": "alexander.shumilov@toptal.com"
		},
		"version": "1.2"
	},
	"servers": [{
			"url": "bonzabiketours-api.herokuapp.com/api/v1",
			"description": "Heroku server - CI"
		},
		{
			"url": "bms.bonzabiketours.alexshumilov.ru:3001/api/v1",
			"description": "Staging server"
		},
		{
			"url": "bms.bonzabiketours.alexshumilov.ru:3002/api/v1",
			"description": "Development server"
		}
	],
	"security": [{
		"api_key": []
	}],
...
	"paths": {
...
		"/users": {
			"get": {
				"tags": [
					"users"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Returns list of users",
				"parameters": [{
						"$ref": "#/components/parameters/limit"
					},
					{
						"$ref": "#/components/parameters/page"
					}
				],
				"responses": {
					"200": {
						"description": "List of users",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/User"
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					}
				}
			},
			"post": {
				"tags": [
					"users"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Creates user",
				"parameters": [],
				"requestBody": {
					"description": "User description",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "User was created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APICreatedEntityResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					}
				}
			}
		},
		"/users/{id}": {
			"parameters": [{
				"$ref": "#/components/parameters/id"
			}],
			"get": {
				"tags": [
					"users"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Returns user description",
				"responses": {
					"200": {
						"description": "User description",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			},
			"put": {
				"tags": [
					"users"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Updates specific user",
				"responses": {
					"200": {
						"description": "Updated user description",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					}
				}
			}
		},
...
		"/settings": {
			"get": {
				"tags": [
					"settings"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Application settings",
				"responses": {
					"200": {
						"description": "Application settings",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/ApplicationSetting"
									}
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"settings"
				],
				"security": [{
					"api_key": []
				}],
				"requestBody": {
					"description": "Application setting description",
					"content": {
						"application/json": {
							"schema": {
								"type": "array",
								"items": {
									"type": "object",
									"properties": {
										"value": {
											"type": "string"
										},
										"key": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				},
				"summary": "Bulk application settings update",
				"responses": {
					"200": {
						"description": "Update application setting",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/reports/monthByMonth": {
			"get": {
				"tags": [
					"reports"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Month by month sales report",
				"responses": {
					"200": {
						"description": "Sales report",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/MonthByMonthSalesReport"
									}
								}
							}
						}
					}
				}
			}
		},
		"/reports/payments": {
			"get": {
				"tags": [
					"reports"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Payments sales report",
				"parameters": [{
						"name": "from",
						"description": "Date start",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"required": true
					},
					{
						"name": "to",
						"description": "Date end",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"required": true
					},
					{
						"name": "city",
						"description": "Tour city",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "Sales report",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/PaymentSalesReport"
									}
								}
							}
						}
					}
				}
			}
		},
		"/reports/payments/export": {
			"get": {
				"tags": [
					"reports"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Payments sales report exported",
				"parameters": [{
						"name": "from",
						"description": "Date start",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"required": true
					},
					{
						"name": "to",
						"description": "Date end",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"required": true
					},
					{
						"name": "city",
						"description": "Tour city",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "Sales report",
						"content": {
							"text/csv": {
								"schema": {
									"$ref": "#/components/schemas/PaymentSalesReportExported"
								}
							}
						}
					}
				}
			}
		},
		"/reports/future": {
			"get": {
				"tags": [
					"reports"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Future sales report",
				"parameters": [{
					"name": "city",
					"description": "Tour city",
					"in": "query",
					"schema": {
						"type": "string"
					}
				}],
				"responses": {
					"200": {
						"description": "Sales report",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/FutureSalesReport"
									}
								}
							}
						}
					}
				}
			}
		},
		"/reports/future/export": {
			"get": {
				"tags": [
					"reports"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Future sales report exported",
				"parameters": [{
					"name": "city",
					"description": "Tour city",
					"in": "query",
					"schema": {
						"type": "string"
					}
				}],
				"responses": {
					"200": {
						"description": "Sales report",
						"content": {
							"text/csv": {
								"schema": {
									"$ref": "#/components/schemas/FutureSalesExported"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"responses": {
...
			"APISuccessResponse": {
				"description": "Operation was completed successfully",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"message": {
									"type": "string"
								}
							}
						}
					}
				}
			},
...
			"APIUnauthorizedResponse": {
				"description": "Authorize in order to access the resource",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"errorType": {
									"type": "string"
								},
								"errorMessage": {
									"type": "string"
								}
							}
						}
					}
				}
			},
			"APINotFoundResponse": {
				"description": "Requested resource was not found",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"errorType": {
									"type": "string"
								},
								"errorMessage": {
									"type": "string"
								}
							}
						}
					}
				}
			},
			"APICreatedEntityResponse": {
				"description": "Returning identifier of created entity",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"id": {
									"type": "integer"
								}
							}
						}
					}
				}
			}
		},
		"parameters": {
			"id": {
				"name": "id",
				"in": "path",
				"description": "Specifies item system identifier",
				"schema": {
					"type": "integer"
				},
				"required": true,
				"style": "simple"
			},
			"limit": {
				"name": "limit",
				"in": "query",
				"description": "Specifies maximum number of fetched items",
				"schema": {
					"type": "integer"
				}
			},
			"page": {
				"name": "page",
				"in": "query",
				"description": "Specifies page of fetched items (offsetting)",
				"schema": {
					"type": "integer"
				}
			}
		},
		"schemas": {
			"MonthByMonthSalesReport": {
				"properties": {
					"id": {
						"type": "integer"
					}
				}
			},
...
			"FutureSalesReport": {
				"properties": {
					"id": {
						"type": "integer"
					}
				}
			},
			"FutureSalesExported": {
				"properties": {
					"id": {
						"type": "integer"
					}
				}
			},
			"PaymentSalesReport": {
				"properties": {
					"id": {
						"type": "integer"
					}
				}
			},
			"PaymentSalesReportExported": {
				"properties": {
					"id": {
						"type": "integer"
					}
				}
			},
...
			"User": {
				"properties": {
					"userId": {
						"type": "integer"
					},
					"lastname": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"accessLevel": {
						"type": "string"
					}
				}
			},
			"ApplicationSetting": {
				"properties": {
					"id": {
						"type": "integer"
					},
					"key": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				}
			},
...
		}
	}
}